backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
django==4.2.4; python_version >= '3.8'
django-cors-headers==4.2.0; python_version >= '3.8'
django-dotenv==1.4.2
django-redis==5.3.0; python_version >= '3.6'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
mysql-connector-python==8.1.0; python_version >= '3.8'
mysqlclient==2.2.0; python_version >= '3.8'
nltk==3.8.1; python_version >= '3.7'
numpy==1.25.2; python_version >= '3.9'
pandas==2.0.3; python_version >= '3.8'
protobuf==4.21.12; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3
redis==5.0.0; python_version >= '3.7'
regex==2023.8.8; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
scikit-learn==1.3.0; python_version >= '3.8'
scipy==1.11.2; python_version < '3.13' and python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sklearn==0.0.post7
sqlparse==0.4.4; python_version >= '3.5'
threadpoolctl==3.2.0; python_version >= '3.8'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.7.1; python_version >= '3.7'
tzdata==2023.3; python_version >= '2'
urllib3==2.0.4; python_version >= '3.7'
